requirements.txt
requests>=2.28.0
Pillow>=9.0.0
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.11
pywin32>=306
opencv-python>=4.8.0
rembg>=2.0.50
numpy>=1.24.0
vosk>=0.3.45
